Update Main.java
I have added some user interaction to the dice game. However, there is a bug where if you get the same value of the dice you do not win the game anymore. Making it impossible to win the game.  I have to challenge myself to write better code to prevent this issue.