Update GamePanel.java
Added update, paintComponent, and run methods. The update and paintComponent methods work hand in hand to move the character and show the character visually. However, there is a problem. The game time is too fast and we must limit the time for it to work properly. Similar to how FPS(Frames per second) works in video games.