Update TileManager.java
I discovered there was no way to have two separate draw methods for the Game. Java is limited to a few things for game design. I wanted my game to be unique with two unique environments. However, it will still be unique when I make the two NPCs that the player can fight.