Update GamePanel.java
Added font and heart objects to the game panel to be able to use them. Added game state to know when the game is running to execute specific methods like "drawPlayerLife." There is now its own draw method in the GamePanel class to help draw hearts with the support of the "drawPlayerLife" method.